// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/CloneAgegroupRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CloneAgegroupRequest = {
    agegroupName: string;
    copyEligibility?: boolean;
    copyRosterSettings?: boolean;
    copyVisualIdentity?: boolean;
    copyFees?: boolean;
    copyDivisions?: boolean;
};